+++ CHANGELOG.md
@@ -4,7 +4,8 @@ All notable changes to this project will be documented in this file.
 ## [1.1] - 2019-11-20
 ### Added
 - added new method metrics::plot_precision_recall for plotting precision recall curve for binary and multi class
-- added matplotlib testing for plot_precision_recall
+- added new method metrics::plot_roc_curve_multi_class for plottingroc curve for multi class
+- added matplotlib testing for plot_precision_recall, plot_roc_curve_binary_class and plot_roc_curve_multi_class
 ### Changed
 - metrics::plot_roc_curve renamed to plot_roc_curve_binary_class
 ### Fixed